+++ test/index_test.js
@@ -327,25 +327,6 @@ tape.test('v5 compile', function(t) {
         t.end();
     });
 
-    t.test('respects options.instructionStringHook', function(assert) {
-        var v5Compiler = compiler('v5', {
-            hooks: {
-                tokenizedInstruction: function(instruction) {
-                    return instruction.replace('{way_name}', '<blink>{way_name}</blink>');
-                }
-            }
-        });
-
-        assert.equal(v5Compiler.compile('en', {
-            maneuver: {
-                type: 'turn',
-                modifier: 'left'
-            },
-            name: 'Way Name'
-        }), 'Turn left onto <blink>Way Name</blink>');
-        assert.end();
-    });
-
     t.test('fixtures match generated instructions', function(assert) {
         // pre-load instructions
         var version = 'v5';